Mejoras a la interfaz del cmp. Tambien se agrega una senial cuando se para de utilizar un slider y se agrega un metodo para obtener el indice actual en un dropdown

+++ coloUi/colouislider.h
@@ -13,7 +13,7 @@ public:
     void setConfiguration(ColoUiConfiguration c);
 
     qint32 getCurrentValue() const {return currentValue;}
-    void setCurrentValue(qint32 v);
+    void setCurrentValue(qreal v);
 
 protected:
     void mousePressEvent(QGraphicsSceneMouseEvent *e);
